Module_10/requirements.txt
aiobreaker==1.2.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiohttp-cors==0.8.1
aiokafka==0.12.0
aiosignal==1.4.0
alembic==1.17.1
amqp==5.3.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.4.0
autopep8==2.3.2
backoff==2.2.1
backports.asyncio.runner==1.2.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
colorlog==6.10.1
confluent-kafka==2.12.2
coverage==7.11.0
croniter==2.0.7
dnspython==2.8.0
docker==7.1.0
exceptiongroup==1.3.0
fastapi==0.120.0
faust-streaming==0.11.3
flower==2.0.1
frozenlist==1.8.0
greenlet==3.2.4
grpcio==1.76.0
# grpcio-tools==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.14.0
idna==3.11
iniconfig==2.3.0
intervaltree==3.1.0
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.3
mode-streaming==0.4.1
motor==3.7.1
multidict==6.7.0
mypy_extensions==1.1.0
opentracing==2.4.0
orjson==3.11.4
packaging==25.0
pika==1.3.2
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pybreaker==1.4.1
pycodestyle==2.14.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pymongo==4.15.3
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
# pywin32==311
redis==5.2.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.44
starlette==0.48.0
terminaltables==3.1.10
# testcontainers==4.13.2
tomli==2.3.0
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
venusian==3.1.0
vine==5.1.0
wcwidth==0.2.14
wrapt==2.0.0
yarl==1.22.0
